refactor: Use a bitmask for opened valves
I'm not sure it was worth the refactor, however I did learn a bit
(hah) about bitmasks, which seems really cool for use-cases
like this where there's a finite list of items and we want to pack
the information into a performant and fast-to-visualise format.

I'm hoping this might be a better way than using sets for part 2?