solve: Day 17 part 2
This one was satisfying to eventually finish! I was struggling
with how to calculate the simulated height since the blocks
wouldn't fall in a nice line, but eventually I ralised it doesn't
matter since I could compare the first ocurrence's rock count
and height with the repeat's and then continue iterating.